Use float32 for better and more stable gpu performance

+++ museval/metrics.py
@@ -526,7 +526,7 @@ def _compute_reference_correlations(reference_sources, filters_len):
     reference_sources = _zeropad(reference_sources, filters_len - 1, axis=2)
     n_fft = int(2**np.ceil(np.log2(nsampl + filters_len - 1.)))
 
-    sf = cupy.asnumpy(cupyx.scipy.fft.rfft(cupy.asarray(reference_sources), n=n_fft, axis=2))
+    sf = cupy.asnumpy(cupyx.scipy.fft.rfft(cupy.asarray(reference_sources, dtype=np.float32), n=n_fft, axis=2))
 
     # compute intercorrelation between sources
     G = np.zeros((nsrc, nsrc, nchan, nchan, filters_len, filters_len))
@@ -538,13 +538,13 @@ def _compute_reference_correlations(reference_sources, filters_len):
         2
     ):
         ssf = cupy.asnumpy(cupyx.scipy.fft.irfft(cupy.asarray(sf[j, c2] * np.conj(sf[i, c1]))))
-
         ss = toeplitz(
             np.hstack((ssf[0], ssf[-1:-filters_len:-1])),
             r=ssf[:filters_len]
         )
         G[j, i, c2, c1] = ss
         G[i, j, c1, c2] = ss.T
+
     return G, sf
 
 
@@ -588,8 +588,8 @@ def _compute_projection_filters(G, sf, estimated_source):
     D = D.reshape(nsrc * nchan * filters_len, nchan)
     G = _reshape_G(G)
 
-    D_gpu = cupy.asarray(D)
-    G_gpu = cupy.asarray(G)
+    D_gpu = cupy.asarray(D, dtype=np.float32)
+    G_gpu = cupy.asarray(G, dtype=np.float32)
 
     # Distortion filters
     try:
